Split the output helpers into their own modules
Move the PyVista side into _output_rendering and the Matplotlib side
into _output_plotting, leaving output.py holding only its four public
functions. The module had grown to the second largest in the package
while mixing two rendering backends that share almost nothing, and the
upcoming CSV export and path plumbing would have pushed it past where
it started. Give each backend its own file so that both stay readable
as the feature work lands.

Give each visual constant to the module that consumes it rather than
gathering them into a shared module. Of the forty constants, thirty six
have exactly one consumer, so a shared module would have been mostly
single-use definitions held at arm's length. Put the four that
genuinely cross a boundary in _output_rendering, which sits lower in
the dependency order, and name that module for rendering rather than
meshing so that owning the styling is honest.

Keep _quality in output.py, which is where its only consumers are.
scripts/regenerate_example_outputs.py re-renders oversized WebPs by
setting ps.output._quality, and that works by rebinding a module
global, so relocating the constant would leave the patch setting an
attribute nothing reads and the re-render running silently at full
quality.
