+++ pterasoftware/output.py
@@ -24,7 +24,6 @@ from __future__ import annotations
 
 import math
 import time
-from typing import NamedTuple
 
 import matplotlib.colors
 import matplotlib.legend_handler
@@ -36,13 +35,11 @@ import webp
 from . import (
     _colormaps,
     _logging,
+    _output_plotting,
+    _output_rendering,
     _parameter_validation,
     _transformations,
     free_flight_unsteady_ring_vortex_lattice_method,
-    geometry,
-)
-from . import operating_point as operating_point_mod
-from . import (
     steady_horseshoe_vortex_lattice_method,
     steady_ring_vortex_lattice_method,
     unsteady_ring_vortex_lattice_method,
@@ -51,29 +48,15 @@ from .movements import free_flight_movement as free_flight_movement_mod
 
 _logger = _logging.get_logger("output")
 
-# Define the colors used by the visualization functions. The color maps and color
-# palettes live in the _colormaps module.
-_wake_vortex_color = "white"
-_panel_color = "chartreuse"
+# Define the Plotter's appearance and the quality of the WebPs it is saved to.
+_plotter_background_color = "black"
 _streamline_color = "orchid"
 _image_surface_opacity = 0.5
-_image_surface_scale = 5.0
-_image_reflection_mute_factor = 0.5
-_image_surface_checker_size = 25
-_image_surface_color_a = np.array([40, 40, 40], dtype=np.uint8)
-_image_surface_color_b = np.array([80, 80, 80], dtype=np.uint8)
-_plotter_background_color = "black"
-_figure_background_color = "None"
-_text_color = (129, 129, 129)
-_text_color_normalized: tuple[float, float, float] = (
-    _text_color[0] / 255,
-    _text_color[1] / 255,
-    _text_color[2] / 255,
-)
-_text_color_surface = (220, 220, 220)
+_text_speed_position = (0.05, 0.075)
 _quality = 75.0
 _window_size = [1024, 768]
 
+# Define the colors of the series in the results plots.
 [
     _alpha_color,
     _beta_color,
@@ -84,25 +67,6 @@ _window_size = [1024, 768]
     _angular_y_color,
     _angular_z_color,
 ] = _colormaps.prism[1:9]
-
-# Set constants for the color maps, scalar bars, and text boxes.
-_color_map_num_sig = 3
-_bar_title_font_size = 30
-_bar_label_font_size = 21
-_bar_width = 0.5
-_bar_position_x = 0.25
-_bar_position_y = 0.05
-_bar_n_labels = 2
-_text_max_position = (0.85, 0.075)
-_text_min_position = (0.85, 0.050)
-_text_speed_position = (0.05, 0.075)
-_text_font_size = 11
-
-# Set the line widths for the results plots. Lines are drawn from thickest to thinnest
-# so that all remain visible even when they overlap.
-_max_line_width = 3.5
-_min_line_width = 1.5
-_legend_line_width = (_max_line_width + _min_line_width) / 2
 
 # Define the camera's view-up direction for free flight visualizations. Earth axes have
 # +z pointing down, so physical up is the -z direction. The free flight visualizations
@@ -260,14 +224,16 @@ def draw(
         qInf__E = draw_operating_point.qInf__E
 
         if is_free_flight:
-            T_pas_GP1_CgP1_to_E_Eo = _get_T_pas_GP1_CgP1_to_E_Eo(draw_operating_point)
+            T_pas_GP1_CgP1_to_E_Eo = _output_rendering.get_T_pas_GP1_CgP1_to_E_Eo(
+                draw_operating_point
+            )
 
         if show_wake_vortices:
-            wake_ring_vortex_surfaces = _get_wake_ring_vortex_surfaces(
+            wake_ring_vortex_surfaces = _output_rendering.get_wake_ring_vortex_surfaces(
                 solver, draw_step
             )
             if T_pas_GP1_CgP1_to_E_Eo is not None:
-                wake_ring_vortex_surfaces = _transform_mesh(
+                wake_ring_vortex_surfaces = _output_rendering.transform_mesh(
                     wake_ring_vortex_surfaces, T_pas_GP1_CgP1_to_E_Eo
                 )
     else:
@@ -276,9 +242,11 @@ def draw(
         qInf__E = draw_operating_point.qInf__E
 
     # Get the Panel surfaces, mapping them into Earth axes for free flight.
-    panel_surfaces = _get_panel_surfaces(airplanes)
+    panel_surfaces = _output_rendering.get_panel_surfaces(airplanes)
     if T_pas_GP1_CgP1_to_E_Eo is not None:
-        panel_surfaces = _transform_mesh(panel_surfaces, T_pas_GP1_CgP1_to_E_Eo)
+        panel_surfaces = _output_rendering.transform_mesh(
+            panel_surfaces, T_pas_GP1_CgP1_to_E_Eo
+        )
 
     T_reflect = draw_operating_point.surfaceReflect_T_act_GP1_CgP1
     image_surface_mesh = None
@@ -296,24 +264,26 @@ def draw(
 
     # Choose the scalar coloring for the Panels, leaving it None to color them
     # uniformly.
-    coloring: _ScalarColoring | None = None
+    coloring: _output_rendering.ScalarColoring | None = None
     if scalar_type in ("induced drag", "side force", "lift"):
-        these_scalars = _get_scalars(airplanes, scalar_type, qInf__E)
-        color_map, c_min, c_max = _choose_color_map(these_scalars)
-        coloring = _ScalarColoring(
+        these_scalars = _output_rendering.get_scalars(airplanes, scalar_type, qInf__E)
+        color_map, c_min, c_max = _output_rendering.choose_color_map(these_scalars)
+        coloring = _output_rendering.ScalarColoring(
             scalars=these_scalars,
             scalar_type=scalar_type,
             min_scalar=float(min(these_scalars)),
             max_scalar=float(max(these_scalars)),
             color_map=color_map,
-            muted_color_map=_mute_colormap(color_map, _image_reflection_mute_factor),
+            muted_color_map=_output_rendering.mute_colormap(
+                color_map, _output_rendering.image_reflection_mute_factor
+            ),
             c_min=c_min,
             c_max=c_max,
         )
 
     # Add the wake, the Panels, and, if an image surface is defined, their reflections.
     # The image surface plane is added later, after the geometry bounds are captured.
-    _add_frame_geometry(
+    _output_rendering.add_frame_geometry(
         plotter,
         panel_surfaces,
         wake_ring_vortex_surfaces,
@@ -382,8 +352,9 @@ def draw(
                                 reflected_point,
                             ),
                             show_edges=True,
-                            color=_mute_color(
-                                _streamline_color, _image_reflection_mute_factor
+                            color=_output_rendering.mute_color(
+                                _streamline_color,
+                                _output_rendering.image_reflection_mute_factor,
                             ),
                             line_width=2,
                             smooth_shading=False,
@@ -401,17 +372,19 @@ def draw(
             # The image surface helper builds the plane from geometry-axis quantities,
             # so it needs geometry-axis bounds. Build the plane there, then map it into
             # Earth axes to match the rendered geometry.
-            geometry_axis_bounds = _get_panel_surfaces(airplanes).bounds
-            image_surface_result = _get_image_surface_mesh_and_texture(
+            geometry_axis_bounds = _output_rendering.get_panel_surfaces(
+                airplanes
+            ).bounds
+            image_surface_result = _output_rendering.get_image_surface_mesh_and_texture(
                 draw_operating_point, geometry_axis_bounds
             )
             assert image_surface_result is not None
             image_surface_mesh, image_surface_texture = image_surface_result
-            image_surface_mesh = _transform_mesh(
+            image_surface_mesh = _output_rendering.transform_mesh(
                 image_surface_mesh, T_pas_GP1_CgP1_to_E_Eo
             )
         else:
-            image_surface_result = _get_image_surface_mesh_and_texture(
+            image_surface_result = _output_rendering.get_image_surface_mesh_and_texture(
                 draw_operating_point, geometry_bounds
             )
             assert image_surface_result is not None
@@ -485,7 +458,7 @@ def draw(
 
     # If saving, take a screenshot and save it as a WebP.
     if save:
-        image = _screenshot_image(plotter)
+        image = _output_rendering.screenshot_image(plotter)
 
         webp.save_image(
             img=image, file_path="draw.webp", lossless=False, quality=_quality
@@ -583,7 +556,7 @@ def animate(
     step_transforms: list[np.ndarray] = []
     if is_free_flight:
         step_transforms = [
-            _get_T_pas_GP1_CgP1_to_E_Eo(steady_problem.operating_point)
+            _output_rendering.get_T_pas_GP1_CgP1_to_E_Eo(steady_problem.operating_point)
             for steady_problem in unsteady_solver.steady_problems
         ]
 
@@ -614,7 +587,7 @@ def animate(
     # time steps and Airplanes. These will be used to set the color map limits.
     if scalar_type is not None:
         for step_id, airplanes in enumerate(step_airplanes):
-            scalars_to_add = _get_scalars(
+            scalars_to_add = _output_rendering.get_scalars(
                 airplanes,
                 scalar_type,
                 unsteady_solver.steady_problems[step_id].operating_point.qInf__E,
@@ -625,8 +598,10 @@ def animate(
         # map and one pair of limits apply to every frame. Mute it once here as well,
         # since building the muted copy walks a 256 entry table and every frame that
         # carries reflected geometry needs it.
-        color_map, c_min, c_max = _choose_color_map(all_scalars)
-        muted_color_map = _mute_colormap(color_map, _image_reflection_mute_factor)
+        color_map, c_min, c_max = _output_rendering.choose_color_map(all_scalars)
+        muted_color_map = _output_rendering.mute_colormap(
+            color_map, _output_rendering.image_reflection_mute_factor
+        )
 
         min_scalar = float(min(all_scalars))
         max_scalar = float(max(all_scalars))
@@ -643,14 +618,18 @@ def animate(
         image_surface_texture,
         T_reflect,
         image_surface_geometry_bounds,
-    ) = _get_animation_image_surface(
+    ) = _output_rendering.get_animation_image_surface(
         unsteady_solver,
         step_airplanes,
         step_transforms,
         is_free_flight,
         show_wake_vortices,
     )
-    animate_text_color = _text_color_surface if T_reflect is not None else _text_color
+    animate_text_color = (
+        _output_rendering.text_color_surface
+        if T_reflect is not None
+        else _output_rendering.text_color
+    )
 
     # For free flight, compute a fixed camera that frames the whole trajectory. The body
     # moves through the scene, so the camera is centered on the trajectory's midpoint
@@ -673,11 +652,11 @@ def animate(
 
         # Map the first and last frames' Panel surfaces into Earth axes. These two
         # frames bound the trajectory, so their combined extent frames the whole glide.
-        first_step_panel_surfaces = _transform_mesh(
-            _get_panel_surfaces(step_airplanes[0]), step_transforms[0]
+        first_step_panel_surfaces = _output_rendering.transform_mesh(
+            _output_rendering.get_panel_surfaces(step_airplanes[0]), step_transforms[0]
         )
-        last_step_panel_surfaces = _transform_mesh(
-            _get_panel_surfaces(step_airplanes[last_step]),
+        last_step_panel_surfaces = _output_rendering.transform_mesh(
+            _output_rendering.get_panel_surfaces(step_airplanes[last_step]),
             step_transforms[last_step],
         )
         airplane_bounds = np.array(first_step_panel_surfaces.bounds, dtype=float)
@@ -705,8 +684,10 @@ def animate(
         framing_meshes = [first_step_panel_surfaces, last_step_panel_surfaces]
         free_flight_clip_meshes = [last_step_panel_surfaces]
         if show_wake_vortices:
-            last_step_wake_surfaces = _transform_mesh(
-                _get_wake_ring_vortex_surfaces(unsteady_solver, last_step),
+            last_step_wake_surfaces = _output_rendering.transform_mesh(
+                _output_rendering.get_wake_ring_vortex_surfaces(
+                    unsteady_solver, last_step
+                ),
                 step_transforms[last_step],
             )
             if last_step_wake_surfaces.n_points > 0:
@@ -717,7 +698,7 @@ def animate(
         # projection is parallel) to the projected extent of that geometry about the
         # focal point. This frames the glide snugly. The user can rescale interactively
         # before the animation is captured.
-        free_flight_parallel_scale = _free_flight_fit_parallel_scale(
+        free_flight_parallel_scale = _output_rendering.free_flight_fit_parallel_scale(
             framing_meshes,
             trajectoryMidpoint_E_Eo,
             _free_flight_view_direction_E,
@@ -729,25 +710,27 @@ def animate(
         plotter.add_text(
             text="Speed: " + str(round(100 * speed)) + "%",
             position=_text_speed_position,
-            font_size=_text_font_size,
+            font_size=_output_rendering.text_font_size,
             viewport=True,
             color=animate_text_color,
         )
 
     # Get the Panel surfaces of the first time step's Airplane(s), mapping them into
     # Earth axes for free flight.
-    panel_surfaces = _get_panel_surfaces(step_airplanes[0])
+    panel_surfaces = _output_rendering.get_panel_surfaces(step_airplanes[0])
     if is_free_flight:
-        panel_surfaces = _transform_mesh(panel_surfaces, step_transforms[0])
+        panel_surfaces = _output_rendering.transform_mesh(
+            panel_surfaces, step_transforms[0]
+        )
 
     # Choose the first time step's scalar coloring, leaving it None to color the Panels
     # uniformly.
-    coloring: _ScalarColoring | None = None
+    coloring: _output_rendering.ScalarColoring | None = None
     if scalar_type is not None and first_results_step == 0:
         assert color_map is not None
         assert muted_color_map is not None
-        coloring = _ScalarColoring(
-            _get_scalars(
+        coloring = _output_rendering.ScalarColoring(
+            _output_rendering.get_scalars(
                 step_airplanes[0],
                 scalar_type,
                 unsteady_solver.steady_problems[0].operating_point.qInf__E,
@@ -763,7 +746,9 @@ def animate(
 
     # Add the first time step's geometry. No wake is passed, since the first time step
     # has not shed one yet.
-    _add_frame_geometry(plotter, panel_surfaces, None, coloring, T_reflect)
+    _output_rendering.add_frame_geometry(
+        plotter, panel_surfaces, None, coloring, T_reflect
+    )
 
     # If an image surface is defined, plot the pre-computed plane, set the camera
     # direction, and fit the camera to the last time step's geometry bounds so the view
@@ -852,7 +837,7 @@ def animate(
         plotter.camera.clipping_range = free_flight_clipping_range
 
     # Start a list to hold a WebP Image of each frame, beginning with this first frame.
-    images = [_screenshot_image(plotter)]
+    images = [_output_rendering.screenshot_image(plotter)]
 
     # Initialize a variable to keep track of the current time step.
     current_step = 1
@@ -865,9 +850,9 @@ def animate(
 
         # Get the Panel surfaces of this time step's Airplane(s), mapping them into
         # Earth axes for free flight.
-        panel_surfaces = _get_panel_surfaces(airplanes)
+        panel_surfaces = _output_rendering.get_panel_surfaces(airplanes)
         if is_free_flight:
-            panel_surfaces = _transform_mesh(
+            panel_surfaces = _output_rendering.transform_mesh(
                 panel_surfaces, step_transforms[current_step]
             )
 
@@ -876,7 +861,7 @@ def animate(
             plotter.add_text(
                 text="Speed: " + str(round(100 * speed)) + "%",
                 position=_text_speed_position,
-                font_size=_text_font_size,
+                font_size=_output_rendering.text_font_size,
                 viewport=True,
                 color=animate_text_color,
             )
@@ -886,11 +871,11 @@ def animate(
         # scene.
         wake_ring_vortex_surfaces = None
         if show_wake_vortices:
-            wake_ring_vortex_surfaces = _get_wake_ring_vortex_surfaces(
+            wake_ring_vortex_surfaces = _output_rendering.get_wake_ring_vortex_surfaces(
                 unsteady_solver, current_step
             )
             if is_free_flight:
-                wake_ring_vortex_surfaces = _transform_mesh(
+                wake_ring_vortex_surfaces = _output_rendering.transform_mesh(
                     wake_ring_vortex_surfaces, step_transforms[current_step]
                 )
 
@@ -900,8 +885,8 @@ def animate(
         if scalar_type is not None and first_results_step <= current_step:
             assert color_map is not None
             assert muted_color_map is not None
-            coloring = _ScalarColoring(
-                _get_scalars(
+            coloring = _output_rendering.ScalarColoring(
+                _output_rendering.get_scalars(
                     airplanes,
                     scalar_type,
                     unsteady_solver.steady_problems[
@@ -918,7 +903,7 @@ def animate(
             )
 
         # Add this time step's geometry.
-        _add_frame_geometry(
+        _output_rendering.add_frame_geometry(
             plotter, panel_surfaces, wake_ring_vortex_surfaces, coloring, T_reflect
         )
 
@@ -943,7 +928,7 @@ def animate(
 
         # If saving, append a WebP Image of this frame to the list of Images.
         if save:
-            images.append(_screenshot_image(plotter))
+            images.append(_output_rendering.screenshot_image(plotter))
 
         # Increment the time step tracker.
         current_step += 1
@@ -1065,7 +1050,7 @@ def plot_results_versus_time(
         # Plot this Airplane's four load figures. The wind axes x and z force components
         # are negated so the series read as induced drag and lift, which point opposite
         # those axes.
-        _plot_time_history(
+        _output_plotting.plot_time_history(
             times,
             [
                 -forces_W[airplane_id, 0],
@@ -1080,7 +1065,7 @@ def plot_results_versus_time(
             save,
             airplane_name_snake + "_forces.png",
         )
-        _plot_time_history(
+        _output_plotting.plot_time_history(
             times,
             [
                 -forceCoefficients_W[airplane_id, 0],
@@ -1099,7 +1084,7 @@ def plot_results_versus_time(
             save,
             airplane_name_snake + "_force_coefficients.png",
         )
-        _plot_time_history(
+        _output_plotting.plot_time_history(
             times,
             [
                 moments_W_Cg[airplane_id, 0],
@@ -1114,7 +1099,7 @@ def plot_results_versus_time(
             save,
             airplane_name_snake + "_moments.png",
         )
-        _plot_time_history(
+        _output_plotting.plot_time_history(
             times,
             [
                 momentCoefficients_W_Cg[airplane_id, 0],
@@ -1168,8 +1153,10 @@ def plot_results_versus_time(
         # Iterate through the time steps and extract each step's state.
         for step, this_operating_point in enumerate(operating_points):
             positions_E_Eo[:, step] = this_operating_point.CgP1_E_Eo
-            velocities_E__E[:, step] = _velocity_E__E_from_operating_point(
-                this_operating_point
+            velocities_E__E[:, step] = (
+                _output_plotting.velocity_E__E_from_operating_point(
+                    this_operating_point
+                )
             )
             anglesDeg_E_to_BP1_izyx[:, step] = this_operating_point.angles_E_to_BP1_izyx
             omegasDeg_BP1__E[:, step] = this_operating_point.omegas_BP1__E
@@ -1181,7 +1168,7 @@ def plot_results_versus_time(
         airplane_name = unsteady_solver.steady_problems[0].airplanes[0].name
         airplane_name_snake = airplane_name.lower().replace(" ", "_")
 
-        _plot_time_history(
+        _output_plotting.plot_time_history(
             state_times,
             [positions_E_Eo[0], positions_E_Eo[1], positions_E_Eo[2]],
             ["X Component", "Y Component", "Z Component"],
@@ -1193,7 +1180,7 @@ def plot_results_versus_time(
             save,
             airplane_name_snake + "_position.png",
         )
-        _plot_time_history(
+        _output_plotting.plot_time_history(
             state_times,
             [velocities_E__E[0], velocities_E__E[1], velocities_E__E[2]],
             ["X Component", "Y Component", "Z Component"],
@@ -1205,7 +1192,7 @@ def plot_results_versus_time(
             save,
             airplane_name_snake + "_velocity.png",
         )
-        _plot_time_history(
+        _output_plotting.plot_time_history(
             state_times,
             [
                 anglesDeg_E_to_BP1_izyx[0],
@@ -1221,7 +1208,7 @@ def plot_results_versus_time(
             save,
             airplane_name_snake + "_orientation.png",
         )
-        _plot_time_history(
+        _output_plotting.plot_time_history(
             state_times,
             [omegasDeg_BP1__E[0], omegasDeg_BP1__E[1], omegasDeg_BP1__E[2]],
             ["Roll Rate", "Pitch Rate", "Yaw Rate"],
@@ -1232,7 +1219,7 @@ def plot_results_versus_time(
             save,
             airplane_name_snake + "_angular_velocity.png",
         )
-        _plot_time_history(
+        _output_plotting.plot_time_history(
             state_times,
             [alphas, betas],
             ["Angle of Attack", "Sideslip Angle"],
@@ -1653,7 +1640,9 @@ def log_results(
         ]:
             CgP1_E_Eo = this_operating_point.CgP1_E_Eo
             angles_E_to_BP1_izyx = this_operating_point.angles_E_to_BP1_izyx
-            velocity_E__E = _velocity_E__E_from_operating_point(this_operating_point)
+            velocity_E__E = _output_plotting.velocity_E__E_from_operating_point(
+                this_operating_point
+            )
             omegas_BP1__E = this_operating_point.omegas_BP1__E
 
             state_component_values = [
@@ -1711,888 +1700,3 @@ def log_results(
                 _logging.indent(2) + f"{beta_label:<{aerodynamic_angle_space}}"
                 f"{this_operating_point.beta:#10.3G} deg"
             )
-
-
-def _velocity_E__E_from_operating_point(
-    this_operating_point: operating_point_mod.OperatingPoint,
-) -> np.ndarray:
-    """Returns the first Airplane's CG velocity (in Earth axes, observed from the Earth
-    frame) for a free flight OperatingPoint.
-
-    The CG velocity is the negative of the freestream velocity, since the freestream (a
-    still airmass) is entirely due to the first Airplane's motion. The OperatingPoint
-    stores the freestream velocity in the first Airplane's geometry axes, so it is
-    rotated into Earth axes here.
-
-    :param this_operating_point: The OperatingPoint whose CG velocity will be returned.
-    :return: A (3,) ndarray of floats representing the first Airplane's CG velocity (in
-        Earth axes, observed from the Earth frame) in meters per second.
-    """
-    vInf_E__E = _transformations.apply_T_to_vectors(
-        this_operating_point.T_pas_GP1_CgP1_to_E_CgP1,
-        this_operating_point.vInf_GP1__E,
-        is_position=False,
-    )
-    return -vInf_E__E
-
-
-def _plot_time_history(
-    times: np.ndarray,
-    series: list[np.ndarray],
-    labels: list[str],
-    colors: list[str],
-    title: str,
-    subtitle: str,
-    y_label: str,
-    save: bool,
-    save_name: str,
-) -> None:
-    """Plots one time-history figure, which is a set of series that share a y-axis and
-    are plotted against time.
-
-    Every figure that plot_results_versus_time produces is drawn through this function,
-    both the per-Airplane load figures and the free flight state-history figures, so all
-    of them share one styling implementation.
-
-    :param times: A (num_steps,) ndarray of floats representing the time, in seconds, at
-        each time step.
-    :param series: A list of (num_steps,) ndarrays of floats, one per line to plot.
-    :param labels: A list of the legend labels, one per series.
-    :param colors: A list of the line colors, one per series.
-    :param title: The figure's title.
-    :param subtitle: A smaller line below the title describing the axes, points, and
-        frames of the plotted quantity. Pass an empty string to omit.
-    :param y_label: The figure's y-axis label.
-    :param save: Set this to True to save the figure as a PNG.
-    :param save_name: The file name to save the figure under if save is True.
-    :return: None
-    """
-    figure, axes = plt.subplots()
-
-    # Remove the plot's top and right spines.
-    axes.spines.right.set_visible(False)
-    axes.spines.top.set_visible(False)
-
-    # Format the plot's spine and label colors.
-    axes.spines.bottom.set_color(_text_color_normalized)
-    axes.spines.left.set_color(_text_color_normalized)
-    axes.xaxis.label.set_color(_text_color_normalized)
-    axes.yaxis.label.set_color(_text_color_normalized)
-
-    # Format the plot's tick colors.
-    axes.tick_params(axis="x", colors=_text_color_normalized)
-    axes.tick_params(axis="y", colors=_text_color_normalized)
-
-    # Format the plot's background colors.
-    figure.patch.set_facecolor(_figure_background_color)
-    axes.set_facecolor(_figure_background_color)
-
-    # Populate the plot. Lines are drawn from thickest to thinnest so that all remain
-    # visible even when the curves overlap.
-    num_series = len(series)
-    widths = np.linspace(_max_line_width, _min_line_width, num_series)
-    for series_id, (this_series, label, color) in enumerate(
-        zip(series, labels, colors)
-    ):
-        axes.plot(
-            times,
-            this_series,
-            label=label,
-            color=color,
-            linewidth=widths[series_id],
-            solid_capstyle="butt",
-        )
-
-    # Name the plot's axis labels, title, and subtitle.
-    axes.set_xlabel("Time (s)", color=_text_color_normalized)
-    axes.set_ylabel(y_label, color=_text_color_normalized)
-    figure.suptitle(title, color=_text_color_normalized)
-    if subtitle:
-        axes.set_title(subtitle, color=_text_color_normalized, fontsize="small")
-
-    # Format the plot's legend.
-    axes.legend(
-        facecolor=_figure_background_color,
-        edgecolor=_figure_background_color,
-        labelcolor=_text_color_normalized,
-        handler_map={
-            plt.Line2D: matplotlib.legend_handler.HandlerLine2D(
-                update_func=lambda h, orig: (
-                    h.update_from(orig),
-                    h.set_linewidth(_legend_line_width),
-                )
-            )
-        },
-    )
-
-    # Save the figure as a PNG if the user wants to do so.
-    if save:
-        figure.savefig(save_name, dpi=300)
-
-
-def _get_panel_surfaces(
-    airplanes: tuple[geometry.airplane.Airplane, ...],
-) -> pv.PolyData:
-    """Returns a PolyData representation of the Wings' Panels' surfaces associated with
-    all the Airplanes in a tuple of Airplanes.
-
-    :param airplanes: The tuple of Airplanes whose Wings' Panels' surfaces will be
-        returned.
-    :return: A PolyData representation of the Airplanes' Wings' Panels' surfaces.
-    """
-    # Initialize empty ndarrays to hold the Panels' vertices and faces.
-    panel_vertices = np.empty((0, 3), dtype=float)
-    panel_faces = np.empty(0, dtype=int)
-
-    # Initialize a variable to keep track of how many Panels have been added thus far.
-    panel_num = 0
-
-    # Increment through the Airplanes' Wing(s).
-    for airplane in airplanes:
-        for wing in airplane.wings:
-            _panels = wing.panels
-            assert _panels is not None
-
-            # Unravel this Wing's ndarray of Panels iterate through it.
-            panels = np.ravel(_panels)
-            for panel in panels:
-                # Arrange this Panel's vertices and faces into ndarrays in the proper
-                # form to represent PolyData surfaces.
-                panel_vertices_to_add = np.vstack(
-                    (
-                        panel.Flpp_GP1_CgP1,
-                        panel.Frpp_GP1_CgP1,
-                        panel.Brpp_GP1_CgP1,
-                        panel.Blpp_GP1_CgP1,
-                    )
-                )
-                panel_face_to_add = np.array(
-                    [
-                        4,
-                        (panel_num * 4),
-                        (panel_num * 4) + 1,
-                        (panel_num * 4) + 2,
-                        (panel_num * 4) + 3,
-                    ],
-                    dtype=int,
-                )
-
-                # Add this Panel's vertices and faces to the ndarray of all vertices and
-                # faces.
-                panel_vertices = np.vstack((panel_vertices, panel_vertices_to_add))
-                panel_faces = np.hstack((panel_faces, panel_face_to_add))
-
-                # Update the number of Panels.
-                panel_num += 1
-
-    # Return the Panels' surfaces.
-    return pv.PolyData(panel_vertices, panel_faces)
-
-
-def _get_image_surface_mesh_and_texture(
-    this_operating_point: operating_point_mod.OperatingPoint,
-    geometry_bounds: tuple[float, float, float, float, float, float],
-) -> tuple[pv.PolyData, pv.Texture] | None:
-    """Returns a PolyData plane mesh and checkerboard Texture representing the image
-    surface, or None if no image surface is defined.
-
-    The plane is centered at the projection of the geometry bounding box center onto the
-    image surface, and sized proportionally to the bounding box diagonal so that it
-    appears large relative to the geometry.
-
-    :param this_operating_point: The OperatingPoint that may define an image surface.
-    :param geometry_bounds: The (xmin, xmax, ymin, ymax, zmin, zmax) bounding box of the
-        geometry used to determine the plane's center and size.
-    :return: A tuple of (PolyData plane mesh, checkerboard Texture) representing the
-        image surface, or None if no image surface is defined.
-    """
-    surface_normal = this_operating_point.surfaceNormal_GP1
-    surface_point = this_operating_point.surfacePoint_GP1_CgP1
-
-    if surface_normal is None or surface_point is None:
-        return None
-
-    # Compute the bounding box center and diagonal length.
-    bounds = np.array(geometry_bounds, dtype=float)
-    bbox_center = np.array(
-        [
-            0.5 * (bounds[0] + bounds[1]),
-            0.5 * (bounds[2] + bounds[3]),
-            0.5 * (bounds[4] + bounds[5]),
-        ],
-        dtype=float,
-    )
-    bbox_diagonal = float(
-        np.linalg.norm(
-            np.array(
-                [
-                    bounds[1] - bounds[0],
-                    bounds[3] - bounds[2],
-                    bounds[5] - bounds[4],
-                ],
-                dtype=float,
-            )
-        )
-    )
-
-    # Project the bounding box center onto the image surface to get the plane's center.
-    offset = np.dot(bbox_center - surface_point, surface_normal)
-    plane_center = bbox_center - offset * surface_normal
-
-    # Size the plane proportionally to the bounding box diagonal.
-    plane_size = _image_surface_scale * bbox_diagonal
-
-    mesh = pv.Plane(
-        center=plane_center,
-        direction=surface_normal,
-        i_size=plane_size,
-        j_size=plane_size,
-    )
-
-    # Build a checkerboard texture image. Each cell is one pixel, so a 25 x 25
-    # checkerboard is a 25 x 25 x 3 RGB image.
-    n = _image_surface_checker_size
-    row = np.arange(n, dtype=int)
-    col = np.arange(n, dtype=int)
-    rr, cc = np.meshgrid(row, col, indexing="ij")
-    is_dark = (rr + cc) % 2 == 0
-    image = np.where(
-        is_dark[:, :, np.newaxis], _image_surface_color_a, _image_surface_color_b
-    )
-    texture = pv.numpy_to_texture(image)
-
-    return mesh, texture
-
-
-def _get_animation_image_surface(
-    unsteady_solver: unsteady_ring_vortex_lattice_method.UnsteadyRingVortexLatticeMethodSolver,
-    step_airplanes: list[tuple[geometry.airplane.Airplane, ...]],
-    step_transforms: list[np.ndarray],
-    is_free_flight: bool,
-    show_wake_vortices: bool,
-) -> tuple[
-    pv.PolyData | None,
-    pv.Texture | None,
-    np.ndarray | None,
-    tuple[float, float, float, float, float, float] | None,
-]:
-    """Builds the image surface plane that an animation reuses for every frame, and
-    returns it alongside the reflection that maps geometry onto its far side.
-
-    The plane is built from the last time step's geometry so that it is large enough to
-    encompass the fully developed wake. It is static across the animation, as is the
-    reflection, so both are computed once here rather than per frame.
-
-    For the standard body-fixed rendering, the plane is sized to the last step's
-    geometry together with its reflected copy, and the wake is included in that extent
-    when it is being shown. For free flight, the plane is fixed in the world while the
-    geometry is rendered in Earth axes, so it is built from the last step's geometry
-    axis bounds and then mapped into Earth axes. The reflection is re-expressed in Earth
-    axes by the same passive transformation, which lets each frame's reflected geometry
-    be built directly from the panels already mapped into Earth axes.
-
-    :param unsteady_solver: The UnsteadyRingVortexLatticeMethodSolver being animated.
-    :param step_airplanes: A list holding each time step's tuple of Airplanes.
-    :param step_transforms: A list holding, per time step, the (4,4) ndarray of floats
-        representing the passive transformation from the first Airplane's geometry axes
-        (relative to its CG) to Earth axes (relative to the Earth origin). It is empty
-        when is_free_flight is False.
-    :param is_free_flight: Set this to True when the solver is a
-        FreeFlightUnsteadyRingVortexLatticeMethodSolver, whose geometry is rendered in
-        Earth axes.
-    :param show_wake_vortices: Set this to True when the wake ring vortices are being
-        shown, which includes them in the extent the plane is sized to.
-    :return: A tuple holding the image surface plane's mesh, its Texture, the (4,4)
-        ndarray of floats representing the active transformation that reflects geometry
-        across the image surface, and the (xmin, xmax, ymin, ymax, zmin, zmax) bounding
-        box to fit the camera to. The bounding box spans the geometry and its reflection
-        but not the much larger plane, so fitting to it keeps the plane from dominating
-        the view. All four are None when no image surface is defined, and the bounding
-        box is also None for free flight, which frames its own camera to the trajectory.
-    """
-    last_step = len(step_airplanes) - 1
-    last_step_operating_point = unsteady_solver.steady_problems[
-        last_step
-    ].operating_point
-    T_reflect = last_step_operating_point.surfaceReflect_T_act_GP1_CgP1
-
-    if T_reflect is None:
-        return None, None, None, None
-
-    last_step_panel_surfaces = _get_panel_surfaces(step_airplanes[last_step])
-
-    if is_free_flight:
-        # Size the plane from the geometry axis bounds, since the image surface helper
-        # builds it from geometry axis quantities, then map it into Earth axes to match
-        # the rendered geometry.
-        image_surface_result = _get_image_surface_mesh_and_texture(
-            last_step_operating_point, last_step_panel_surfaces.bounds
-        )
-        assert image_surface_result is not None
-        image_surface_mesh, image_surface_texture = image_surface_result
-
-        T_pas_last = step_transforms[last_step]
-        image_surface_mesh = _transform_mesh(image_surface_mesh, T_pas_last)
-        T_reflect = T_pas_last @ T_reflect @ _transformations.invert_T_pas(T_pas_last)
-
-        # Free flight frames its camera to the whole trajectory, so it needs no geometry
-        # bounding box from here.
-        return image_surface_mesh, image_surface_texture, T_reflect, None
-
-    # Size the plane to the last step's geometry together with its reflected copy, so
-    # that it spans both sides of the surface.
-    reflected_last_step_panel_surfaces = _reflect_mesh(
-        last_step_panel_surfaces, T_reflect
-    )
-    if show_wake_vortices:
-        last_step_wake_surfaces = _get_wake_ring_vortex_surfaces(
-            unsteady_solver, last_step
-        )
-        combined = (
-            last_step_panel_surfaces.merge(last_step_wake_surfaces)
-            .merge(reflected_last_step_panel_surfaces)
-            .merge(_reflect_mesh(last_step_wake_surfaces, T_reflect))
-        )
-    else:
-        combined = last_step_panel_surfaces.merge(reflected_last_step_panel_surfaces)
-
-    image_surface_geometry_bounds = combined.bounds
-    image_surface_result = _get_image_surface_mesh_and_texture(
-        last_step_operating_point, image_surface_geometry_bounds
-    )
-    assert image_surface_result is not None
-    image_surface_mesh, image_surface_texture = image_surface_result
-
-    return (
-        image_surface_mesh,
-        image_surface_texture,
-        T_reflect,
-        image_surface_geometry_bounds,
-    )
-
-
-def _get_T_pas_GP1_CgP1_to_E_Eo(
-    this_operating_point: operating_point_mod.OperatingPoint,
-) -> np.ndarray:
-    """Returns the passive transformation from the first Airplane's geometry axes,
-    relative to the first Airplane's CG, to Earth axes, relative to the Earth origin.
-
-    Free flight visualizations render each time step's geometry in its true Earth-frame
-    position and orientation, so the body translates and rotates through the scene. This
-    transformation chains the per-step geometry-to-Earth rotation with the translation
-    from the first Airplane's CG to the Earth origin.
-
-    :param this_operating_point: The OperatingPoint whose Earth-frame position and
-        orientation define the transformation.
-    :return: A (4,4) ndarray of floats representing the passive transformation from the
-        first Airplane's geometry axes, relative to the first Airplane's CG, to Earth
-        axes, relative to the Earth origin.
-    """
-    # Translate from Earth axes relative to the first Airplane's CG to Earth axes
-    # relative to the Earth origin. For a passive translation, the parameter is the
-    # position of the final reference point (the Earth origin) relative to the initial
-    # one (the first Airplane's CG), which is the negative of the CG's position relative
-    # to the Earth origin.
-    T_pas_E_CgP1_to_E_Eo = _transformations.generate_trans_T(
-        translations=-this_operating_point.CgP1_E_Eo,
-        passive=True,
-    )
-
-    # Chain geometry-to-Earth (relative to the CG) with the CG-to-origin translation.
-    return _transformations.compose_T_pas(
-        this_operating_point.T_pas_GP1_CgP1_to_E_CgP1,
-        T_pas_E_CgP1_to_E_Eo,
-    )
-
-
-def _transform_mesh(
-    mesh: pv.PolyData,
-    T_pas: np.ndarray,
-) -> pv.PolyData:
-    """Returns a copy of a PolyData mesh with its points mapped through a passive
-    transformation.
-
-    :param mesh: The PolyData mesh to transform.
-    :param T_pas: A (4,4) ndarray of floats representing the passive transformation to
-        apply to the mesh's points.
-    :return: A new PolyData mesh with all points mapped through the transformation.
-    """
-    transformed = mesh.copy()
-    transformed.points = _transformations.apply_T_to_vectors(
-        T_pas,
-        mesh.points,
-        is_position=True,
-    )
-    return transformed
-
-
-def _reflect_mesh(
-    mesh: pv.PolyData,
-    T_reflect: np.ndarray,
-) -> pv.PolyData:
-    """Returns a copy of a PolyData mesh with its points reflected across the image
-    surface.
-
-    :param mesh: The PolyData mesh to reflect.
-    :param T_reflect: A (4,4) ndarray of floats representing the active reflection
-        transformation matrix (in the first Airplane's geometry axes, relative to the
-        first Airplane's CG).
-    :return: A new PolyData mesh with all points reflected across the image surface.
-    """
-    return _transform_mesh(mesh, T_reflect)
-
-
-def _free_flight_fit_parallel_scale(
-    meshes: list[pv.PolyData],
-    focalPoint_E_Eo: np.ndarray,
-    viewDirection_E: np.ndarray,
-    viewUp_E: np.ndarray,
-    margin: float = 1.15,
-) -> float:
-    """Returns a parallel-projection scale that frames a set of meshes about a focal
-    point.
-
-    The scale is half the viewport's height in world units (the convention for a
-    parallel projection). It is sized to the largest projection of the meshes' bounding-
-    box corners onto the camera's screen-right and screen-up axes, measured from the
-    focal point, so everything stays in view regardless of the viewport's aspect ratio.
-    A margin leaves a little space around the geometry.
-
-    :param meshes: The PolyData meshes to frame.
-    :param focalPoint_E_Eo: A (3,) ndarray of floats locating the camera's focal point
-        (in Earth axes, relative to the Earth origin). Extents are measured from this
-        point, since it projects to the center of the viewport.
-    :param viewDirection_E: A (3,) ndarray of floats giving the offset from the focal
-        point to the camera position (in Earth axes). The camera looks back along it.
-    :param viewUp_E: A (3,) ndarray of floats giving the camera's up direction (in Earth
-        axes).
-    :param margin: A factor (at least 1.0) by which to pad the fitted scale. The default
-        is 1.15.
-    :return: The parallel-projection scale.
-    """
-    # Build the camera's screen right and up axes in Earth axes. The camera looks from
-    # its position back toward the focal point, i.e. along the negative view direction.
-    # The up axis is the supplied up made orthogonal to that look direction.
-    lookDirection_E = -viewDirection_E
-    lookDirection_E = lookDirection_E / np.linalg.norm(lookDirection_E)
-    upDirection_E = viewUp_E - np.dot(viewUp_E, lookDirection_E) * lookDirection_E
-    upDirection_E = upDirection_E / np.linalg.norm(upDirection_E)
-    rightDirection_E = np.cross(lookDirection_E, upDirection_E)
-    rightDirection_E = rightDirection_E / np.linalg.norm(rightDirection_E)
-
-    # Collect every mesh's eight bounding-box corners, measured from the focal point.
-    corners_E_Eo: list[list[float]] = []
-    for mesh in meshes:
-        bounds = np.array(mesh.bounds, dtype=float)
-        corners_E_Eo.extend(
-            [float(x), float(y), float(z)]
-            for x in bounds[0:2]
-            for y in bounds[2:4]
-            for z in bounds[4:6]
-        )
-    corners_E = np.array(corners_E_Eo) - focalPoint_E_Eo
-
-    # Each matrix-vector product is a batch of dot products, one per corner: the (N, 3)
-    # array of corners times a (3,) screen axis gives an (N,) array whose entries are
-    # each corner's signed distance from the focal point along that axis (the axes are
-    # unit vectors, so the dot product is the scalar projection). Take the largest
-    # magnitude along either axis so neither screen dimension is clipped.
-    half_extent_right = float(np.abs(corners_E @ rightDirection_E).max())
-    half_extent_up = float(np.abs(corners_E @ upDirection_E).max())
-    return margin * max(half_extent_right, half_extent_up)
-
-
-def _mute_color(
-    color: str | tuple[float, ...],
-    factor: float,
-) -> tuple[float, float, float]:
-    """Returns a muted version of a color by linearly interpolating it toward middle
-    gray.
-
-    :param color: Any color that PyVista can parse (name, hex string, RGB tuple, etc.).
-    :param factor: The muting factor in [0, 1]. 0 means no change, 1 means fully gray.
-    :return: A (R, G, B) tuple of floats in [0, 1].
-    """
-    rgb = np.array(pv.Color(color).float_rgb)
-    gray = np.full(3, 0.5)
-    muted = rgb + factor * (gray - rgb)
-    return float(muted[0]), float(muted[1]), float(muted[2])
-
-
-def _mute_colormap(
-    cmap: matplotlib.colors.Colormap,
-    factor: float,
-) -> matplotlib.colors.ListedColormap:
-    """Returns a muted version of a colormap by linearly interpolating each color toward
-    middle gray.
-
-    :param cmap: The colormap to mute.
-    :param factor: The muting factor in [0, 1]. 0 means no change, 1 means fully gray.
-    :return: A ListedColormap with muted colors.
-    """
-    colors = cmap(np.linspace(0, 1, 256))
-    gray = 0.5
-    colors[:, :3] = colors[:, :3] + factor * (gray - colors[:, :3])
-    return matplotlib.colors.ListedColormap(colors)
-
-
-def _get_wake_ring_vortex_surfaces(
-    solver: unsteady_ring_vortex_lattice_method.UnsteadyRingVortexLatticeMethodSolver,
-    step: int,
-) -> pv.PolyData:
-    """Returns the PolyData representation of the surfaces of an
-    UnsteadyRingVortexLatticeMethodSolver's wake ring vortices at a given time step.
-
-    :param solver: The UnsteadyRingVortexLatticeMethodSolver with the wake ring vortices
-        to process.
-    :param step: The time step at which to process the wake ring vortices.
-    :return: The PolyData representation of the wake ring vortices.
-    """
-    num_wake_ring_vortices = solver.list_num_wake_vortices[step]
-    stackFrwrvp_GP1_CgP1 = solver.listStackFrwrvp_GP1_CgP1[step]
-    stackFlwrvp_GP1_CgP1 = solver.listStackFlwrvp_GP1_CgP1[step]
-    stackBlwrvp_GP1_CgP1 = solver.listStackBlwrvp_GP1_CgP1[step]
-    stackBrwrvp_GP1_CgP1 = solver.listStackBrwrvp_GP1_CgP1[step]
-
-    # Initialize empty ndarrays to hold each wake ring vortex's vertices and face.
-    wake_ring_vortex_vertices = np.zeros((0, 3), dtype=float)
-    wake_ring_vortex_faces = np.zeros(0, dtype=int)
-
-    for wake_ring_vortex_num in range(num_wake_ring_vortices):
-        Frwrvp_GP1_CgP1 = stackFrwrvp_GP1_CgP1[wake_ring_vortex_num]
-        Flwrvp_GP1_CgP1 = stackFlwrvp_GP1_CgP1[wake_ring_vortex_num]
-        Blwrvp_GP1_CgP1 = stackBlwrvp_GP1_CgP1[wake_ring_vortex_num]
-        Brwrvp_GP1_CgP1 = stackBrwrvp_GP1_CgP1[wake_ring_vortex_num]
-
-        wake_ring_vortex_vertices_to_add = np.vstack(
-            (
-                Flwrvp_GP1_CgP1,
-                Frwrvp_GP1_CgP1,
-                Brwrvp_GP1_CgP1,
-                Blwrvp_GP1_CgP1,
-            )
-        )
-        wake_ring_vortex_face_to_add = np.array(
-            [
-                4,
-                (wake_ring_vortex_num * 4),
-                (wake_ring_vortex_num * 4) + 1,
-                (wake_ring_vortex_num * 4) + 2,
-                (wake_ring_vortex_num * 4) + 3,
-            ],
-            dtype=int,
-        )
-
-        # Stack this wake ring vortex's vertices and faces to the ndarrays of all wake
-        # ring vortices' vertices and faces.
-        wake_ring_vortex_vertices = np.vstack(
-            (wake_ring_vortex_vertices, wake_ring_vortex_vertices_to_add)
-        )
-        wake_ring_vortex_faces = np.hstack(
-            (wake_ring_vortex_faces, wake_ring_vortex_face_to_add)
-        )
-
-    # Return the wake ring vortex surfaces.
-    return pv.PolyData(wake_ring_vortex_vertices, wake_ring_vortex_faces)
-
-
-def _get_scalars(
-    airplanes: tuple[geometry.airplane.Airplane, ...],
-    scalar_type: str,
-    qInf__E: float,
-) -> np.ndarray:
-    """Returns the load coefficient values from a SteadyProblem's Airplanes' Wings'
-    Panels.
-
-    :param airplanes: The tuple of Airplanes with the scalars to return.
-    :param scalar_type: Determines which load coefficient to return as scalars. Can be
-        "induced drag", "side force", or "lift", which respectively use each Panel's
-        induced drag, side force, and lift coefficient.
-    :param qInf__E: The current freestream dynamic pressure experienced by this
-        SteadyProblem's Airplane(s) (observed in the Earth frame). The units are in
-        Pascals.
-    :return: A (N,) ndarray of floats representing the N Panels' load coefficients.
-    """
-    scalars = np.empty(0, dtype=float)
-
-    # Iterate through the Airplanes' Wings.
-    for airplane in airplanes:
-        for wing in airplane.wings:
-            _panels = wing.panels
-            assert _panels is not None
-
-            # Unravel this Wing's ndarray of Panels iterate through them.
-            these_panels = np.ravel(_panels)
-            for this_panel in these_panels:
-
-                # Stack this Panel's scalars.
-                if scalar_type == "induced drag":
-                    this_induced_drag_coefficient = (
-                        -this_panel.forces_W[0] / qInf__E / this_panel.area
-                    )
-
-                    scalars = np.hstack((scalars, this_induced_drag_coefficient))
-
-                if scalar_type == "side force":
-                    this_side_force_coefficient = (
-                        this_panel.forces_W[1] / qInf__E / this_panel.area
-                    )
-
-                    scalars = np.hstack((scalars, this_side_force_coefficient))
-
-                if scalar_type == "lift":
-                    this_lift_coefficient = (
-                        -this_panel.forces_W[2] / qInf__E / this_panel.area
-                    )
-
-                    scalars = np.hstack((scalars, this_lift_coefficient))
-
-    # Return the resulting ndarray of scalars.
-    return scalars
-
-
-def _choose_color_map(
-    scalars: np.ndarray,
-) -> tuple[matplotlib.colors.Colormap, float, float]:
-    """Chooses the color map to color a set of scalars with and finds the limits to
-    apply to it.
-
-    Scalars that keep one sign get a sequential color map, since their values run in a
-    single direction. Scalars that change sign get a diverging color map with limits
-    placed symmetrically about zero, so the map's midpoint marks where the scalar
-    changes sign. In both cases the limits are held within a fixed number of standard
-    deviations of the mean so that a few outlying Panels cannot flatten the contrast
-    across the rest of the geometry.
-
-    :param scalars: A (N,) ndarray of floats representing the scalar value at each of
-        the N Panels being colored.
-    :return: A tuple holding the color map, the lower limit to apply to it, and the
-        upper limit to apply to it.
-    """
-    if np.sign(np.min(scalars)) == np.sign(np.max(scalars)):
-        color_map: matplotlib.colors.Colormap = _colormaps.sequential_color_map
-        c_min = max(
-            float(np.mean(scalars)) - _color_map_num_sig * float(np.std(scalars)),
-            float(np.min(scalars)),
-        )
-        c_max = min(
-            float(np.mean(scalars)) + _color_map_num_sig * float(np.std(scalars)),
-            float(np.max(scalars)),
-        )
-    else:
-        color_map = _colormaps.diverging_color_map
-        c_min = -_color_map_num_sig * float(np.std(scalars))
-        c_max = _color_map_num_sig * float(np.std(scalars))
-
-    return color_map, c_min, c_max
-
-
-# The webp package re-exports Pillow's Image module under the name webp.Image, so the
-# Image class itself is webp.Image.Image. Naming the class through webp keeps this
-# module off Pillow, which Ptera Software only depends on transitively through webp.
-def _screenshot_image(plotter: pv.Plotter) -> webp.Image.Image:
-    """Takes a screenshot of a Plotter and returns it as an Image.
-
-    :param plotter: The Plotter to capture.
-    :return: The captured frame as an Image with a transparent background.
-    """
-    return webp.Image.fromarray(
-        np.array(
-            plotter.screenshot(
-                filename=None,
-                transparent_background=True,
-                return_img=True,
-            )
-        )
-    )
-
-
-def _plot_scalars(
-    plotter: pv.Plotter,
-    these_scalars: np.ndarray,
-    scalar_type: str,
-    min_scalar: float,
-    max_scalar: float,
-    color_map: matplotlib.colors.Colormap,
-    c_min: float,
-    c_max: float,
-    panel_surfaces: pv.PolyData,
-    text_color: tuple[int, int, int] = _text_color,
-) -> None:
-    """Plots a scalar bar, the surfaces of a set of Panels with particular scalars, and
-    labels for the minimum and maximum scalar values.
-
-    :param plotter: The Plotter used for visualization.
-    :param these_scalars: A (N,) ndarray of floats representing the N Panels' load
-        coefficients.
-    :param scalar_type: Which load coefficient is represented by the scalars. Can be
-        "induced drag", "side force", or "lift".
-    :param min_scalar: Minimum scalar value, which is displayed as text on the Plotter.
-    :param max_scalar: Maximum scalar value, which is displayed as text on the Plotter.
-    :param color_map: The color map to use for scalar visualization.
-    :param c_min: Lower bound for the color map scaling.
-    :param c_max: Upper bound for the color map scaling.
-    :param panel_surfaces: PolyData representing the Panels' surfaces.
-    :param text_color: The color used for the scalar bar and label text. The default is
-        _text_color.
-    :return: None
-    """
-    scalar_bar_args = dict(
-        title=scalar_type.title() + " Coefficient",
-        title_font_size=_bar_title_font_size,
-        label_font_size=_bar_label_font_size,
-        width=_bar_width,
-        position_x=_bar_position_x,
-        position_y=_bar_position_y,
-        n_labels=_bar_n_labels,
-        fmt="%#.3G",
-        color=text_color,
-    )
-    plotter.add_mesh(
-        panel_surfaces,
-        show_edges=True,
-        cmap=color_map,
-        clim=[c_min, c_max],
-        scalars=these_scalars,
-        smooth_shading=False,
-        scalar_bar_args=scalar_bar_args,  # type: ignore[arg-type]
-    )
-
-    plotter.add_text(
-        text=f"Max: {max_scalar:#.3G}",
-        position=_text_max_position,
-        font_size=_text_font_size,
-        viewport=True,
-        color=text_color,
-    )
-    plotter.add_text(
-        text=f"Min: {min_scalar:#.3G}",
-        position=_text_min_position,
-        font_size=_text_font_size,
-        viewport=True,
-        color=text_color,
-    )
-
-
-class _ScalarColoring(NamedTuple):
-    """Everything needed to color one frame's Panels by a scalar, and to color their
-    reflected copies to match.
-
-    Only the scalars themselves change from frame to frame during an animation. The
-    remaining fields are constant across the whole animation, because the color map and
-    its limits are chosen once from the scalars of every time step so that a given color
-    means the same value in every frame. muted_color_map is carried here, rather than
-    derived where it is used, because building it walks a 256 entry table and it would
-    otherwise be rebuilt on every frame.
-    """
-
-    scalars: np.ndarray
-    scalar_type: str
-    min_scalar: float
-    max_scalar: float
-    color_map: matplotlib.colors.Colormap
-    muted_color_map: matplotlib.colors.Colormap
-    c_min: float
-    c_max: float
-
-
-def _add_frame_geometry(
-    plotter: pv.Plotter,
-    panel_surfaces: pv.PolyData,
-    wake_surfaces: pv.PolyData | None,
-    coloring: _ScalarColoring | None,
-    T_reflect: np.ndarray | None,
-) -> None:
-    """Adds one frame's geometry to a Plotter, which is the wake, the Panels, and, when
-    an image surface is defined, a muted reflected copy of both.
-
-    A single drawing and every frame of an animation are the same scene, so both are
-    assembled here. The meshes arrive already mapped into whichever axes they are being
-    rendered in, so this function does not need to know whether it is drawing a free
-    flight body in Earth axes or a body-fixed one in geometry axes.
-
-    The image surface plane itself is not added here. A drawing sizes its plane from the
-    bounds of the geometry added above, so it cannot be added until afterward, and it
-    must in any case be added last so it renders over the geometry it is blended with.
-
-    :param plotter: The Plotter to add the geometry to.
-    :param panel_surfaces: The PolyData representation of the frame's Panel surfaces.
-    :param wake_surfaces: The PolyData representation of the frame's wake ring vortex
-        surfaces, or None to omit the wake.
-    :param coloring: The scalar coloring to apply to the Panels, or None to color them
-        uniformly.
-    :param T_reflect: A (4,4) ndarray of floats representing the active transformation
-        that reflects geometry across the image surface, or None when no image surface
-        is defined, in which case no reflected geometry is added.
-    :return: None
-    """
-    # Add the wake ring vortex surfaces if they are being shown.
-    if wake_surfaces is not None:
-        plotter.add_mesh(
-            wake_surfaces,
-            show_edges=True,
-            smooth_shading=False,
-            color=_wake_vortex_color,
-        )
-
-    # Plot the Panels either with scalar coloring or with a uniform color.
-    if coloring is not None:
-        _plot_scalars(
-            plotter,
-            coloring.scalars,
-            coloring.scalar_type,
-            coloring.min_scalar,
-            coloring.max_scalar,
-            coloring.color_map,
-            coloring.c_min,
-            coloring.c_max,
-            panel_surfaces,
-            text_color=_text_color_surface if T_reflect is not None else _text_color,
-        )
-    else:
-        plotter.add_mesh(
-            panel_surfaces,
-            show_edges=True,
-            color=_panel_color,
-            smooth_shading=False,
-        )
-
-    if T_reflect is None:
-        return
-
-    # An image surface is defined, so add the reflected geometry. It is muted toward
-    # gray so that it reads as a reflection rather than as more geometry.
-    mute = _image_reflection_mute_factor
-    muted_edge_color = _mute_color("black", mute)
-
-    # Add reflected Panel surfaces with muted coloring.
-    reflected_panel_surfaces = _reflect_mesh(panel_surfaces, T_reflect)
-    if coloring is not None:
-        plotter.add_mesh(
-            reflected_panel_surfaces,
-            show_edges=True,
-            edge_color=muted_edge_color,
-            cmap=coloring.muted_color_map,
-            clim=[coloring.c_min, coloring.c_max],
-            scalars=coloring.scalars,
-            smooth_shading=False,
-            show_scalar_bar=False,
-        )
-    else:
-        plotter.add_mesh(
-            reflected_panel_surfaces,
-            show_edges=True,
-            edge_color=muted_edge_color,
-            color=_mute_color(_panel_color, mute),
-            smooth_shading=False,
-        )
-
-    # Add reflected wake ring vortex surfaces if they are being shown.
-    if wake_surfaces is not None:
-        plotter.add_mesh(
-            _reflect_mesh(wake_surfaces, T_reflect),
-            show_edges=True,
-            edge_color=muted_edge_color,
-            smooth_shading=False,
-            color=_mute_color(_wake_vortex_color, mute),
-        )
